pyproject.toml
[project]
name = "nsftbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.22.0",
    "numpy==1.23.5",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "tflite-runtime==2.14.0",
]
